# php/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-BookingHealthcare API - Swagger",
        "description": "E-BookingHealthcare API with Laravel",
        "contact": {
            "name": "Nguyen Chinh",
            "url": "https://github.com/ngchinhdev/",
            "email": "chinhnguyennn24@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/medicine-categories": {
            "get": {
                "tags": [
                    "Medicine category Routes"
                ],
                "summary": "Get all Medicine category",
                "operationId": "be5c195249e0f0149df427c75112d2e9",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
        "/api/v1/medicine-categories/{id}": {
            "get": {
                "tags": [
                    "Medicine category Routes"
                ],
                "summary": "Get one Medicine category",
                "operationId": "c1b2249b2912f0b11ebe9ddd4815debe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
        "/api/v1/medicine-categories/add": {
            "post": {
                "tags": [
                    "Medicine category Routes"
                ],
                "summary": "Add a Medicine category",
                "operationId": "da2fc57d7f567178c7c8900033de0ca1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New MedicineCategory"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
        "/api/v1/medicine-categories/update/{id}": {
            "put": {
                "tags": [
                    "Medicine category Routes"
                ],
                "summary": "Update Medicine category",
                "operationId": "1d83481ef0e0bf126fdfdfd8c930a8da",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "object id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New MedicineCategory"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
        "/api/v1/medicine-categories/delete/{id}": {
            "delete": {
                "tags": [
                    "Medicine category Routes"
                ],
                "summary": "Delete Medicine category",
                "operationId": "4f694c1105cc385f78497c9e7ad2a875",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "object id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response"
                    }
                }
            }
        },
...
    },
...
}